refactor tests and benchmarking to unify dependency and remove duplicated structures/code
