+++ frame/crowdloan-rewards/src/mocks.rs
@@ -1,12 +1,21 @@
-use crate::{self as pallet_crowdloan_rewards};
-use frame_support::{construct_runtime, parameter_types, traits::Everything};
+use crate::{
+	self as pallet_crowdloan_rewards,
+	models::{EcdsaSignature, EthereumAddress, Proof, RemoteAccount},
+};
+use codec::Encode;
+use frame_support::{
+	construct_runtime, dispatch::DispatchResultWithPostInfo, parameter_types, traits::Everything,
+};
 use frame_system as system;
-use sp_core::H256;
+use sp_core::{ed25519, keccak_256, Pair, H256};
 use sp_runtime::{
 	traits::{BlakeTwo256, ConvertInto, IdentityLookup},
 	AccountId32, Perbill,
 };
 use system::EnsureRoot;
+
+pub type RelayKey = ed25519::Pair;
+pub type EthKey = libsecp256k1::SecretKey;
 
 pub type BlockNumber = u32;
 pub type AccountId = AccountId32;
@@ -85,7 +94,6 @@ impl pallet_crowdloan_rewards::Config for Test {
 	type VestingStep = VestingStep;
 	type Prefix = Prefix;
 	type AdminOrigin = EnsureRoot<AccountId>;
-	type AssociationOrigin = EnsureRoot<AccountId>;
 	type WeightInfo = ();
 }
 
@@ -122,4 +130,113 @@ impl ExtBuilder {
 			.unwrap();
 		t.into()
 	}
+}
+
+#[derive(Clone)]
+pub enum ClaimKey {
+	Relay(RelayKey),
+	Eth(EthKey),
+}
+
+impl ClaimKey {
+	pub fn as_remote_public(&self) -> RemoteAccount<RelayChainAccountId> {
+		match self {
+			ClaimKey::Relay(relay_account) =>
+				RemoteAccount::RelayChain(relay_account.public().into()),
+			ClaimKey::Eth(ethereum_account) =>
+				RemoteAccount::Ethereum(ethereum_address(ethereum_account)),
+		}
+	}
+	pub fn proof(self, reward_account: AccountId32) -> Proof<[u8; 32]> {
+		match self {
+			ClaimKey::Relay(relay) => relay_proof(&relay, reward_account),
+			ClaimKey::Eth(eth) => ethereum_proof(&eth, reward_account),
+		}
+	}
+	pub fn claim(&self, reward_account: AccountId) -> DispatchResultWithPostInfo {
+		CrowdloanRewards::claim(Origin::signed(reward_account))
+	}
+	pub fn associate(&self, reward_account: AccountId) -> DispatchResultWithPostInfo {
+		let proof = match self {
+			ClaimKey::Relay(relay_account) => relay_proof(relay_account, reward_account.clone()),
+			ClaimKey::Eth(ethereum_account) =>
+				ethereum_proof(ethereum_account, reward_account.clone()),
+		};
+		CrowdloanRewards::associate(Origin::none(), reward_account, proof)
+	}
+}
+
+pub fn relay_proof(
+	relay_account: &RelayKey,
+	reward_account: AccountId,
+) -> Proof<RelayChainAccountId> {
+	let mut msg = PROOF_PREFIX.to_vec();
+	msg.append(&mut reward_account.using_encoded(|x| hex::encode(x).as_bytes().to_vec()));
+	Proof::RelayChain(relay_account.public().into(), relay_account.sign(&msg).into())
+}
+
+pub fn ethereum_proof(
+	ethereum_account: &EthKey,
+	reward_account: AccountId,
+) -> Proof<RelayChainAccountId> {
+	let msg = keccak_256(
+		&pallet_crowdloan_rewards::ethereum_signable_message(
+			PROOF_PREFIX,
+			&reward_account.using_encoded(|x| hex::encode(x).as_bytes().to_vec()),
+		)[..],
+	);
+	let (sig, recovery_id) =
+		libsecp256k1::sign(&libsecp256k1::Message::parse(&msg), ethereum_account);
+	let mut r = [0_u8; 65];
+	r[0..64].copy_from_slice(&sig.serialize()[..]);
+	r[64] = recovery_id.serialize();
+	Proof::Ethereum(EcdsaSignature(r))
+}
+
+pub fn ethereum_public(secret: &EthKey) -> libsecp256k1::PublicKey {
+	libsecp256k1::PublicKey::from_secret_key(secret)
+}
+
+pub fn ethereum_address(secret: &EthKey) -> EthereumAddress {
+	let mut res = EthereumAddress::default();
+	res.0
+		.copy_from_slice(&keccak_256(&ethereum_public(secret).serialize()[1..65])[12..]);
+	res
+}
+
+pub fn relay_generate(count: u64) -> Vec<(AccountId, ClaimKey)> {
+	let seed: u128 = 12345678901234567890123456789012;
+	(0..count)
+		.map(|i| {
+			let account_id =
+				[[0_u8; 16], (&(i as u128 + 1)).to_le_bytes()].concat().try_into().unwrap();
+			(
+				AccountId::new(account_id),
+				ClaimKey::Relay(ed25519::Pair::from_seed(&keccak_256(
+					&[(&(seed + i as u128)).to_le_bytes(), (&(seed + i as u128)).to_le_bytes()]
+						.concat(),
+				))),
+			)
+		})
+		.collect()
+}
+
+pub fn ethereum_generate(count: u64) -> Vec<(AccountId, ClaimKey)> {
+	(0..count)
+		.map(|i| {
+			let account_id =
+				[(&(i as u128 + 1)).to_le_bytes(), [0_u8; 16]].concat().try_into().unwrap();
+			(
+				AccountId::new(account_id),
+				ClaimKey::Eth(EthKey::parse(&keccak_256(&i.to_le_bytes())).unwrap()),
+			)
+		})
+		.collect()
+}
+
+pub fn generate_accounts(count: u64) -> Vec<(AccountId, ClaimKey)> {
+	let mut x = relay_generate(count / 2);
+	let mut y = ethereum_generate(count / 2);
+	x.append(&mut y);
+	x
 }
